services/orchestrator/pyproject.toml
[project]
name = "sona-orchestrator"
version = "0.1.0"
description = "Sona Agent Orchestrator — Gemini Live API, ADK tools, multimodal input"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.133.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "google-adk>=1.10",
    "google-genai",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "sympy>=1.14.0",
    "websockets>=15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
